% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19tr_w}
\alias{c19tr_w}
\alias{c19tr_l}
\title{Covid-19 Turkey Dataset}
\format{
An object of class \code{data.frame} with 14 rows and 5 columns.

An object of class \code{data.frame} with 56 rows and 7 columns.
}
\source{
\code{c19tr}: \url{https://twitter.com/drfahrettinkoca}
}
\description{
\code{c19tr_w}: Data for Turkey (Wide format)

\code{c19tr_l}: Data for Turkey (Long format)
}
\details{
Covid-19 datasets for turkey.
}
\keyword{datasets}
