% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19trw}
\alias{c19trw}
\alias{c19trl}
\title{Covid-19 Turkey Dataset}
\format{
An object of class \code{data.frame} with 20 rows and 7 columns.

An object of class \code{data.frame} with 120 rows and 7 columns.
}
\source{
\url{https://twitter.com/drfahrettinkoca}
}
\description{
\code{c19trw}: Wide format

\code{c19trl}: Long format
}
\details{
Covid-19 datasets for turkey.
}
\keyword{datasets}
