% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{c19tr}
\alias{c19tr}
\alias{c19trl}
\alias{c19trp}
\title{Covid-19 Turkey Dataset}
\format{
An object of class \code{data.frame} with 36 rows and 7 columns.

An object of class \code{data.frame} with 216 rows and 7 columns.

An object of class \code{data.frame} with 162 rows and 4 columns.
}
\source{
\url{https://twitter.com/drfahrettinkoca}
}
\description{
\code{c19tr}: Wide format

\code{c19trl}: Long format

\code{c19trp}: Dataset for provinces
}
\details{
Covid-19 datasets for turkey.
}
\keyword{datasets}
